+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Initial release
 - Basic support for Makefile
+- Basic support for CMakeLists.txt